cleanup batcj
