+++ crates/cubecl-matmul/src/components/batch/matmul/mod.rs
@@ -1,0 +1,5 @@
+mod config;
+mod matmul;
+mod partition;
+mod partitioner;
+mod setup;
