+++ crates/cubecl-matmul/src/components/batch/matmul/partition/mod.rs
@@ -1,0 +1,3 @@
+mod matmul;
+
+pub use matmul::*;
