+++ crates/cubecl-matmul/src/components/batch/matmul/partitioner.rs
