+++ crates/cubecl-matmul/src/components/batch/mod.rs
@@ -1,12 +1,8 @@
-pub mod partitioned_batch_matmul;
-
 mod base;
-mod cube_dispatch;
-mod partition_batch_matmul;
+mod entry_point;
+mod matmul;
 mod shared;
 
 pub use base::*;
-pub use cube_dispatch::*;
-pub use partition_batch_matmul::*;
 
 pub use shared::*;
