+++ crates/cubecl-matmul/src/components/stage/matmul/mod.rs
@@ -1,5 +1,7 @@
 mod config;
 mod partition;
-mod partitioned_stage;
+mod partitioned_matmul;
+mod plane_partitioned;
+mod unit_partitioned;
 
 pub use config::*;
