+++ crates/cubecl-matmul/src/components/stage/matmul/partition/fragments.rs
@@ -1,7 +1,7 @@
 use crate::components::MatmulPrecision;
 use crate::components::global::AccumulatorLoader;
 use crate::components::stage::StageConfig;
-use crate::components::stage::matmul::config::CommonStageConfig;
+use crate::components::stage::matmul::config::PartitionedStageConfig;
 use crate::components::tile::TileMatmul;
 use cubecl::prelude::*;
 use cubecl_core as cubecl;
@@ -15,7 +15,7 @@ pub struct Accumulators<MP: MatmulPrecision, TMM: TileMatmul<MP>> {
 
 #[cube]
 impl<MP: MatmulPrecision, TMM: TileMatmul<MP>> Accumulators<MP, TMM> {
-    pub fn new(#[comptime] config: CommonStageConfig<TMM::Config>) -> Accumulators<MP, TMM> {
+    pub fn new(#[comptime] config: PartitionedStageConfig<TMM::Config>) -> Accumulators<MP, TMM> {
         let partition_size = config.tiling_scheme().partition_size;
         let mut accumulators = Sequence::new();
 
@@ -29,7 +29,7 @@ impl<MP: MatmulPrecision, TMM: TileMatmul<MP>> Accumulators<MP, TMM> {
         }
     }
 
-    pub fn zero(&mut self, #[comptime] config: CommonStageConfig<TMM::Config>) {
+    pub fn zero(&mut self, #[comptime] config: PartitionedStageConfig<TMM::Config>) {
         #[unroll]
         for i in 0..comptime![config.tiling_scheme().partition_size.mn()] {
             TMM::zero_accumulator(self.sequence.index_mut(i), config.tile_config());
@@ -39,7 +39,7 @@ impl<MP: MatmulPrecision, TMM: TileMatmul<MP>> Accumulators<MP, TMM> {
     pub fn fill<L: AccumulatorLoader<MP>>(
         &mut self,
         loader: &mut L,
-        #[comptime] config: CommonStageConfig<TMM::Config>,
+        #[comptime] config: PartitionedStageConfig<TMM::Config>,
     ) {
         #[unroll]
         for i in 0..comptime![config.tiling_scheme().partition_size.mn()] {
@@ -52,7 +52,7 @@ impl<MP: MatmulPrecision, TMM: TileMatmul<MP>> Accumulators<MP, TMM> {
         this: &Accumulators<MP, TMM>,
         #[comptime] i: u32,
         #[comptime] j: u32,
-        #[comptime] config: CommonStageConfig<TMM::Config>,
+        #[comptime] config: PartitionedStageConfig<TMM::Config>,
     ) -> &TMM::Accumulator {
         this.sequence.index(comptime!(
             i * config.tiling_scheme().tiles_in_stage_partition_n() + j
@@ -63,7 +63,7 @@ impl<MP: MatmulPrecision, TMM: TileMatmul<MP>> Accumulators<MP, TMM> {
         this: &mut Accumulators<MP, TMM>,
         #[comptime] i: u32,
         #[comptime] j: u32,
-        #[comptime] config: CommonStageConfig<TMM::Config>,
+        #[comptime] config: PartitionedStageConfig<TMM::Config>,
     ) -> &mut TMM::Accumulator {
         this.sequence.index_mut(comptime!(
             i * config.tiling_scheme().tiles_in_stage_partition_n() + j
