+++ crates/cubecl-matmul/src/components/stage/matmul/partition/mod.rs
