--- crates/cubecl-matmul/src/components/stage/matmul/partitioned_stage/mod.rs
@@ -1,3 +1,0 @@
-mod matmul;
-mod plane_partitioned;
-mod unit_partitioned;
